requirements.txt
discord.py
python-valve==0.2.1
